refactor: rename & refactor files in theme folder
