+++ src/index.tsx
@@ -69,7 +69,7 @@ export { default as GlobalStyles } from './utils/GlobalStyles';
 export { default as useSelectableTableRows } from './utils/useSelectableTableRows';
 export { default as usePaginatedTable } from './utils/usePaginatedTable';
 export { default as useSortableColumns } from './utils/useSortableColumns';
-export { default as defaultTheme } from './themes/default';
+export { Theme } from './theme';
 export { default as ThemeProvider } from './utils/ThemeProvider';
 export { default as useTheme } from './utils/useTheme';
