+++ src/theme/index.ts
@@ -1,0 +1,3 @@
+export { default } from './theme';
+export { default as icons } from './icons';
+export * from './theme';
